backend 1/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
docopt==0.6.2
executing==2.2.1
fastapi==0.129.2
fastjsonschema==2.21.2
filelock==3.20.0
flatbuffers==25.12.19
fpdf==1.7.2
fsspec==2025.12.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
lazy_loader==0.4
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
networkx==3.6.1
numba==0.64.0
numpy==2.3.5
onnxruntime==1.24.2
packaging==26.0
pandocfilters==1.5.1
parso==0.8.6
pickleshare==0.7.5
pillow==12.1.1
pipreqs==0.5.0
platformdirs==4.9.2
pooch==1.9.0
postgrest==2.28.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.0
pure_eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.0
PyJWT==2.11.0
PyMatting==1.1.15
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyzmq==27.1.0
realtime==2.28.0
referencing==0.37.0
rembg==2.0.72
requests==2.32.5
resend==2.22.0
rich==14.3.3
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.1
setuptools==70.2.0
six==1.17.0
soupsieve==2.8.3
stack-data==0.6.3
starlette==0.52.1
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
sympy==1.14.0
tenacity==9.1.4
tifffile==2026.2.24
tinycss2==1.4.0
torch==2.10.0+cpu
torchaudio==2.10.0+cpu
torchvision==0.25.0+cpu
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.6.0
webencodings==0.5.1
websockets==15.0.1
yarg==0.1.9
yarl==1.22.0
zstandard==0.25.0
